html/parsing/html5lib/html-tree/adoption02.dat

#data
<b>1<i>2<p>3</b>4
#errors
1
2
3
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "1"
|       <i>
|         "2"
|     <i>
|       <p>
|         <b>
|           "3"
|         "4"

#data
<a><div><style></style><address><a>
#errors
1
2
3
4
5
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|         <style>
|       <address>
|         <a>
|         <a>
